+++ requirements.txt
@@ -1,1 +1,4 @@
+click
 pytest
+pytest-cov
+pytest-pretty
